requirements.txt
fastapi
uvicorn[standard]
aiofiles
python-multipart
pillow
pytesseract
openai
google-genai
sqlalchemy
databases
python-dotenv
jinja2
passlib
bcrypt==4.1.2